% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ateCAL.R
\name{ateCAL}
\alias{ateCAL}
\title{ateCAL: reweight experimental results to match external sample}
\usage{
ateCAL(
  y,
  a,
  X,
  targetMoments,
  treatProb = NULL,
  nuismod = c("rlm", "rf"),
  estimator = c("ACW", "CW"),
  glmnet_lamchoice = "lambda.min",
  glmnet_alpha = 1,
  separateMus = TRUE,
  noi = FALSE
)
}
\arguments{
\item{y}{outcome vector [length n]}

\item{a}{treatment vector [length n]}

\item{X}{covariate matrix [n X k]}

\item{targetMoments}{k-vector of target moments (means in overall sample for
generalization, means in target sample for transportation}

\item{treatProb}{treatment probability, non-null results in no pscore fit}

\item{nuismod}{one of c("rlm", "rf") : choose how to fit nuisance functions
(cross-fit) for  outcome models}

\item{estimator}{one of c("CW", "ACW"). Calibration weights (CW), which fit a set of balancing weights that reweights the sample to match target sample moments. ACW augments this with an outcome model (default)}

\item{glmnet_lamchoice}{choice of lambda (shrinkage parameter) for regularized linear regressions. Only relevant when nuismod == "rlm"}

\item{glmnet_alpha}{in [0, 1], choice of alpha in glmnet. 1 (default) corresponds with L1 regularization (LASSO) and 0 corresponds with L2 regularization (ridge), while intermediate values correspond with a mix of the two (elastic net)}

\item{separateMus}{boolean for whether to fit separate outcome models for each treatment group or a single pooled model. The former is recommended and is the default, but a pooled model may be fit when data is scarce / computation is burdensome.}

\item{noi}{boolean for printing marginal means and causal contrasts table (it gets returned anyway). On by default.}
}
\value{
list containing treatment effects table , nuisance function estimates, and influence function values
}
\description{
Generalization or transportation estimator with method of
 moments weighting. Imputes counterfactual outcome for each observation i under each treament a as
\deqn{Y^a = \omega \frac{A = a}{\pi^a (X)} (Y - \hat{\mu}^a(X)) + \mu^a(X) }
Where \eqn{\omega} is solved for using entropy balancing for balance with target moments.
}
\examples{
df = selDGP(n = 10000)
id = with(df, s == 1)
X = df$X[id == 1,]; y = df$y[id == 1]; a = df$a[id == 1]
Xtar = colMeans(df$X[!id, ])
ateCAL(y, a, X, targetMoments = Xtar, estimator = "ACW") \%>\% summary()
}
