% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfmTest.R
\name{dfmTest}
\alias{dfmTest}
\title{Omnibus test of systematic treatment effect heterogeneity (Ding, Feller, Miratrix 2018)}
\usage{
dfmTest(y, a, X)
}
\arguments{
\item{y}{outcome vector}

\item{a}{treament vector}

\item{X}{covariate matrix}
}
\value{
vector aith chi-squared statistic and p-value for systematic heterogeneity
}
\description{
Omnibus test of treatment effect heterogeneity along specified dimensions X. Implemented by regressing individual treatment effect on covariates and
joint test of covariate coefficients = 0
}
\references{
Ding, P., A. Feller, and L. Miratrix. (2019): “Decomposing Treatment Effect Variation,” Journal of the American Statistical Association, 114, 304–17.
}
