% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ateGT.R
\name{plot.ategt}
\alias{plot.ategt}
\title{SMD balance plot method - plots balance across covariates and gains from reweighting}
\usage{
\method{plot}{ategt}(fit, X)
}
\arguments{
\item{fit}{ategt object}

\item{covariate}{matrix}
}
\description{
SMD balance plot method - plots balance across covariates and gains from reweighting
}
\examples{
sim = selDGP(selF = \(x) -3 * x[5] + x[2] + 3 * x[10])
fit = with(sim, ateGT(y = y, a = a, X = X, s = s))
fit \%>\% summary
mean(sim$tau)
plot(fit, sim$X)
}
