% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{polySieveM}
\alias{polySieveM}
\title{prepare X matrix with flexible interactions and polynomials}
\usage{
polySieveM(
  dat,
  dummies = NULL,
  continuouses = NULL,
  corr_cut = 0.9,
  k = 2,
  m = 2,
  raw = F
)
}
\arguments{
\item{dat}{data table / dataframe}

\item{dummies}{Names of dummy vars (null by default - auto-detected)}

\item{continuouses}{Names of continuous vars to modify fn form (null by default - auto detected)}

\item{corr_cut}{cutoff for correlation threshold to drop one of the vars (default = 0.9)}

\item{k}{order of interactions: defaults to pairwise interactions}

\item{m}{order of functions: defaults to quadratic functions}

\item{raw}{raw or orthogonal bases}
}
\value{
data.frame with base terms and interactions + basis
}
\description{
Prepare matrix with polynomial basis expansion and/or interactions.
}
