% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aipw.R
\name{summary.aipw}
\alias{summary.aipw}
\title{Summary method - prints treatment effects and marginal means}
\usage{
\method{summary}{aipw}(obj)
}
\arguments{
\item{fitted}{aipw object}
}
\description{
Summary method - prints treatment effects and marginal means
}
