% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ateCAL.R
\name{summary.atecal}
\alias{summary.atecal}
\title{summary method - prints treatment effects and marginal means}
\usage{
\method{summary}{atecal}(obj)
}
\arguments{
\item{fitted}{atecal object}
}
\description{
summary method - prints treatment effects and marginal means
}
