% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ateGT.R
\name{summary.ategt}
\alias{summary.ategt}
\title{summary method - prints treatment effects and marginal means}
\usage{
\method{summary}{ategt}(obj)
}
\arguments{
\item{fitted}{ategt object}
}
\description{
summary method - prints treatment effects and marginal means
}
